custom_gcodes: multple changes:
Prepare Tram-Z for upcoming merge (changes to upstream API)
Only enable LOAD/UNLOAD if extruder_type is set to either "prusa" or "skelestruder"
Create a FINISH_MOVES gcode that waits for all prior moves before continuing.
Create a TIMED_GCODE that can execute a gcode with at a delayed time.  Note that when entering a gcode whitespace must be replaced with underscores.