hd44780: noise compensation fix
This patch attempts to deal with noise by invalidating one line of the old frame_buffer every 5 seconds.  This forces a refresh on the device, which should clear garbage charcters generated by noise.

It also reverts the previous attempt to change the init sequence, as it had little to no affect on noise compensation.