filament_sensor: experimental pat9125 support
This commit add experimental support for the PAT9125EL laser tracking sensor, along with support for basic switch style filament sensors such as endstops and the Finda.  For those looking to test the pat9125 with a Prusa i3 MK3 it is STRONGLY recommended that you are either connected to the Raspberry Pi via the header or your have flashed your Atmega32u2 with updated firmware.

Patches incuded in this commit:
- Initial implementation of filament_sensor.py and pat9125.py
- Addition of pat9125.c for the MCU side implementation
- Update to /src/Makefile to include pat9125.c
- Update to stepper.c and stepper.h to allow access to stepper_get_position()
- Update to virtual_sdcard.py to notify Octoprint of the current file position while pause, as well the addition of an is_active() function
-  Update to toolhead.py to include the RECOVER_POSITION gcode along with the necessary functionality to support capturing/restore after a pause (this is subject to change)
- Updates to example-extras.cfg and G-Codes.md for supporting documentation

This commit relies on the following previous patches not yet in the upstream repo:
- update to src/avr/i2c.c to include i2c read support and the ability to set the rate.
- the addition of i2ccmds.h and updates to i2ccmds.c to include i2cdev_oid_lookup(), i2cdev_read(), and i2cdev_write().  Also included is a fix in command_i2c_read that affects AVR based MCUs.